Fix Windows build compatibility (MSVC cl.exe and clang-cl) (#99)

+++ test/test_functions.hpp
@@ -17,7 +17,7 @@ struct AssociatedLegendre {
 
 struct SphericalHarmonic {
   static auto evaluate(int l, int m, double theta, double phi) {
-    double fac_ratio  = detail::factorial(l-m);
+    double fac_ratio  = static_cast<double>(detail::factorial(l-m));
            fac_ratio /= detail::factorial(l+m);
     double prefactor  = (2.0 * l + 1.0) / (4.0 * M_PI);
            prefactor *= fac_ratio;
